move index.js into root
