+++ index.js
@@ -1,0 +1,7 @@
+/**
+ * @template {{}} ApiType
+ * @typedef {import('./lib/types.js').ClientApi<ApiType>} ClientApi
+ */
+
+export { createClient } from './lib/client.js'
+export { createServer } from './lib/server.js'
