--- lib/index.js
@@ -1,7 +1,0 @@
-/**
- * @template {{}} ApiType
- * @typedef {import('./types.js').ClientApi<ApiType>} ClientApi
- */
-
-export { createClient } from './client.js'
-export { createServer } from './server.js'
